// src/app/shared/models/region-stats.model.ts
export interface RegionStats {
  id: number;
  name: string;
  code: string;
  geographicalZone: string;
  economicDescription: string;
  taxIncentives: string;
  infrastructure: string;
  approvedInvestmentServices: number;
  approvedCollaborationServices: number;
  approvedTouristServices: number;
  totalApprovedServices: number;
}